[Update] 알람화면의 pickerView가 나타나는 방식을 바꿨습니다.
