+++ FeedPet/FeedPet/MyPage/AlarmSettingView/AlarmSettingViewController.swift
@@ -14,22 +14,16 @@ import UserNotifications
 class AlarmSettingViewController: UIViewController,UITableViewDataSource,UITableViewDelegate,CustomCellUpdater {
     
     var isGrantedNotificationAccess = false
-    
-    let formatter:DateFormatter = DateFormatter() // "aa hh:mm"
-    let dateFormatterAMPM = DateFormatter()
-    let dateFormatterHour:DateFormatter = DateFormatter() // hh
-    let dateFormatterMinute:DateFormatter = DateFormatter() // mm
-    let locKo = Locale(identifier: "ko")
+        
     var cellIdentificationNumber = 0
     
     
     @IBOutlet weak var tableView: UITableView!
-    @IBOutlet weak var timePickerViewOut: UIDatePicker!
-    @IBOutlet weak var timePickConfirmBtnOut: UIButton!
+
     
     override func viewDidLoad() {
         super.viewDidLoad()
-        
+
         if let alarmOnOffData = UserDefaults.standard.dictionary(forKey: userDefaultsName.alarmOnOff){
             AlarmService.shared.switchOnOff = alarmOnOffData as! [String : Bool]
         }
@@ -51,8 +45,7 @@ class AlarmSettingViewController: UIViewController,UITableViewDataSource,UITable
         }
         
         
-        formatter.locale = locKo
-        timePickerViewOut.locale = locKo
+        
         
 //        let center = UNUserNotificationCenter.current() // 로컬 및 원격 Notification에 대한 권한 요청
 //        center.delegate = self
@@ -108,7 +101,7 @@ class AlarmSettingViewController: UIViewController,UITableViewDataSource,UITable
         }else{
             let switchCell:AlarmSwitchCell = tableView.dequeueReusableCell(withIdentifier: "AlarmSwitchCell", for: indexPath) as! AlarmSwitchCell
             
-                switchCell.delegate = self
+            switchCell.delegate = self
             
             if indexPath.row == 1{
                 switchCell.mealTimeLb.text = AlarmService.shared.mealTime["morning"]!
@@ -130,34 +123,32 @@ class AlarmSettingViewController: UIViewController,UITableViewDataSource,UITable
     
     func tableView(_ tableView: UITableView, didSelectRowAt indexPath: IndexPath) {
         
+        let alarmMealTimePickerView:AlarmMealTimePickerViewController = storyboard?.instantiateViewController(withIdentifier: "AlarmMealTimePickerViewController") as! AlarmMealTimePickerViewController
+        
         if indexPath.row == 1{
             cellIdentificationNumber = 1
+            alarmMealTimePickerView.cellIdentificationNumber = 1
             
-            tableView.isHidden = true
-            
-            timePickerViewOut.isHidden = false
-            timePickConfirmBtnOut.isHidden = false
+            alarmMealTimePickerView.delegate = self
             
             print("morning",AlarmService.shared.mealTime["morning"]!)
-            
+            self.present(alarmMealTimePickerView, animated: true, completion: nil)
         }else if indexPath.row == 2{
             cellIdentificationNumber = 2
-           
-            tableView.isHidden = true
+            alarmMealTimePickerView.cellIdentificationNumber = 2
             
-            timePickerViewOut.isHidden = false
-            timePickConfirmBtnOut.isHidden = false
+            alarmMealTimePickerView.delegate = self
             
-             print("lunch",AlarmService.shared.mealTime["lunch"]!)
+            print("lunch",AlarmService.shared.mealTime["lunch"]!)
+            self.present(alarmMealTimePickerView, animated: true, completion: nil)
         }else if indexPath.row == 3{
             cellIdentificationNumber = 3
+            alarmMealTimePickerView.cellIdentificationNumber = 3
             
-            tableView.isHidden = true
-            
-            timePickerViewOut.isHidden = false
-            timePickConfirmBtnOut.isHidden = false
+            alarmMealTimePickerView.delegate = self
             
             print("dinner",AlarmService.shared.mealTime["dinner"]!)
+            self.present(alarmMealTimePickerView, animated: true, completion: nil)
         }
     }
     
@@ -173,9 +164,6 @@ class AlarmSettingViewController: UIViewController,UITableViewDataSource,UITable
     
     func setMealTimeAlarmNotification() { //알람 Notification을 셋팅 해주는 함수입니다
         
-        dateFormatterHour.dateFormat = "HH" //포메터를 시간으로 세팅을 해줍니다
-        
-        dateFormatterMinute.dateFormat = "mm" //포메터를 분으로 세팅을 해줍니다
         
         var notificationDateComponents = DateComponents()
         
@@ -316,131 +304,128 @@ class AlarmSettingViewController: UIViewController,UITableViewDataSource,UITable
         UserDefaults.standard.set(AlarmService.shared.switchOnOff, forKey: userDefaultsName.alarmOnOff)
         self.navigationController?.popViewController(animated: true)
     }
-    @IBAction func saveBtnAction(_ sender: UIBarButtonItem) {
-      
     
-    }
     
 
     
     @IBAction func timePickConfirmBtnAction(_ sender: UIButton) {
         
-        tableView.isHidden = false
-        
-        formatter.locale = locKo // 포메터 지역을 한국(Ko)으로 설정
-        formatter.dateFormat = "aa hh:mm" //ex) 오전 11:00
-        
-        dateFormatterAMPM.dateFormat = "aa"
-        dateFormatterHour.dateFormat = "hh"
-        dateFormatterMinute.dateFormat = "mm"
-        
-        let dateString = formatter.string(from: timePickerViewOut.date)
-        let hourString = dateFormatterHour.string(from: timePickerViewOut.date)
-        let minuteStr = dateFormatterMinute.string(from: timePickerViewOut.date)
-        let AMPMStr = dateFormatterAMPM.string(from: timePickerViewOut.date)
-        print("!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!",hourString)
-        if cellIdentificationNumber == 1{
-            
-            AlarmService.shared.mealTime["morning"] = dateString          //셀의 레이블에 담을 피커뷰의 시간을 담습니다
-            
-            AlarmService.shared.mealTimeAMPM["morning"] = AMPMStr
-            
-            if AlarmService.shared.mealTimeAMPM["morning"] == "AM" || AlarmService.shared.mealTimeAMPM["morning"] == "오전" {
-                if hourString == "12"{
-                    AlarmService.shared.mealTimeHour["morning"] = 0 //오전12시일경우 매칭할 시간이 0시가 됩니다 그래서 0으로 바꿔줍니다
-                }else{
-                AlarmService.shared.mealTimeHour["morning"] = Int(hourString)
-                }
-            }else if AlarmService.shared.mealTimeAMPM["morning"] == "PM" || AlarmService.shared.mealTimeAMPM["morning"] == "오후"{
-                if hourString == "12"{
-                    AlarmService.shared.mealTimeHour["morning"] = Int(hourString)
-                }else{
-                    AlarmService.shared.mealTimeHour["morning"] = Int(hourString)! + 12 //오후 1시일 경우 매칭할 숫자가 13시입니다 그래서 12를 더해 담습니다
-                }
-                
-            }
-            
-            AlarmService.shared.mealTimeMinute["morning"] = Int(minuteStr) //피커뷰의 분을 담습니다
-            
-            AlarmService.shared.switchOnOff["total"] = true
-            AlarmService.shared.switchOnOff["morning"] = true
-            
-            tableView.reloadData()
-            
-        }else if cellIdentificationNumber == 2{
-            
-            AlarmService.shared.mealTime["lunch"] = dateString
-            
-            
-            AlarmService.shared.mealTimeAMPM["lunch"] = AMPMStr
-            
-            if AlarmService.shared.mealTimeAMPM["lunch"] == "AM" || AlarmService.shared.mealTimeAMPM["lunch"] == "오전" {
-                if hourString == "12"{
-                    AlarmService.shared.mealTimeHour["lunch"] = 0 //오전12시일경우 매칭할 시간이 0시가 됩니다 그래서 0으로 바꿔줍니다
-                }else{
-                    AlarmService.shared.mealTimeHour["lunch"] = Int(hourString)
-                }
-            }else if AlarmService.shared.mealTimeAMPM["lunch"] == "PM" || AlarmService.shared.mealTimeAMPM["lunch"] == "오후"{
-                if hourString == "12"{
-                    AlarmService.shared.mealTimeHour["lunch"] = Int(hourString)
-                }else{
-                    AlarmService.shared.mealTimeHour["lunch"] = Int(hourString)! + 12 //오후 1시일 경우 매칭할 숫자가 13시입니다 그래서 12를 더해 담습니다
-                }
-                
-            }
-            
-            AlarmService.shared.mealTimeMinute["lunch"] = Int(minuteStr)
-            
-            AlarmService.shared.switchOnOff["total"] = true
-            AlarmService.shared.switchOnOff["lunch"] = true
-            
-            tableView.reloadData()
-            
-        }else if cellIdentificationNumber == 3{
-            
-            AlarmService.shared.mealTime["dinner"] = dateString
-            
-            AlarmService.shared.mealTimeAMPM["dinner"] = AMPMStr
-            
-            if AlarmService.shared.mealTimeAMPM["dinner"] == "AM" || AlarmService.shared.mealTimeAMPM["dinner"] == "오전" {
-                if hourString == "12"{
-                    AlarmService.shared.mealTimeHour["dinner"] = 0 //오전12시일경우 매칭할 시간이 0시가 됩니다 그래서 0으로 바꿔줍니다
-                }else{
-                    AlarmService.shared.mealTimeHour["dinner"] = Int(hourString)
-                }
-            }else if AlarmService.shared.mealTimeAMPM["dinner"] == "PM" || AlarmService.shared.mealTimeAMPM["dinner"] == "오후"{
-                if hourString == "12"{
-                    AlarmService.shared.mealTimeHour["dinner"] = Int(hourString)
-                }else{
-                    AlarmService.shared.mealTimeHour["dinner"] = Int(hourString)! + 12 //오후 1시일 경우 매칭할 숫자가 13시입니다 그래서 12를 더해 담습니다
-                }
-                
-            }
-            
-            AlarmService.shared.mealTimeMinute["dinner"] = Int(minuteStr)
-            
-            AlarmService.shared.switchOnOff["total"] = true
-            AlarmService.shared.switchOnOff["dinner"] = true
-            
-            tableView.reloadData()
-            
-        }
-        
-        UserDefaults.standard.setValue(AlarmService.shared.mealTime, forKey: userDefaultsName.mealTime)
-        UserDefaults.standard.setValue(AlarmService.shared.mealTimeHour, forKey: userDefaultsName.mealTimeHour)
-        UserDefaults.standard.setValue(AlarmService.shared.mealTimeMinute, forKey: userDefaultsName.mealTimeMinute)
-        UserDefaults.standard.setValue(AlarmService.shared.mealTimeAMPM, forKey: userDefaultsName.mealTimeAMPM)
-        UserDefaults.standard.setValue(AlarmService.shared.switchOnOff, forKey: userDefaultsName.alarmOnOff)
-        
-        print(UserDefaults.standard.dictionary(forKey: userDefaultsName.alarmOnOff) ?? "알람OnOff값이 없음")
-        print(AlarmService.shared.mealTimeAMPM)
-        
-        setMealTimeAlarmNotification()
-
-        timePickerViewOut.isHidden = true
-        timePickConfirmBtnOut.isHidden = true
-        
-       
+//        tableView.isHidden = false
+//
+//        formatter.locale = locKo // 포메터 지역을 한국(Ko)으로 설정
+//        formatter.dateFormat = "aa hh:mm" //ex) 오전 11:00
+//
+//        dateFormatterAMPM.dateFormat = "aa"
+//        dateFormatterHour.dateFormat = "hh"
+//        dateFormatterMinute.dateFormat = "mm"
+//
+//        let dateString = formatter.string(from: timePickerViewOut.date)
+//        let hourString = dateFormatterHour.string(from: timePickerViewOut.date)
+//        let minuteStr = dateFormatterMinute.string(from: timePickerViewOut.date)
+//        let AMPMStr = dateFormatterAMPM.string(from: timePickerViewOut.date)
+//        print("!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!",hourString)
+//        if cellIdentificationNumber == 1{
+//
+//            AlarmService.shared.mealTime["morning"] = dateString          //셀의 레이블에 담을 피커뷰의 시간을 담습니다
+//
+//            AlarmService.shared.mealTimeAMPM["morning"] = AMPMStr
+//
+//            if AlarmService.shared.mealTimeAMPM["morning"] == "AM" || AlarmService.shared.mealTimeAMPM["morning"] == "오전" {
+//                if hourString == "12"{
+//                    AlarmService.shared.mealTimeHour["morning"] = 0 //오전12시일경우 매칭할 시간이 0시가 됩니다 그래서 0으로 바꿔줍니다
+//                }else{
+//                AlarmService.shared.mealTimeHour["morning"] = Int(hourString)
+//                }
+//            }else if AlarmService.shared.mealTimeAMPM["morning"] == "PM" || AlarmService.shared.mealTimeAMPM["morning"] == "오후"{
+//                if hourString == "12"{
+//                    AlarmService.shared.mealTimeHour["morning"] = Int(hourString)
+//                }else{
+//                    AlarmService.shared.mealTimeHour["morning"] = Int(hourString)! + 12 //오후 1시일 경우 매칭할 숫자가 13시입니다 그래서 12를 더해 담습니다
+//                }
+//
+//            }
+//
+//            AlarmService.shared.mealTimeMinute["morning"] = Int(minuteStr) //피커뷰의 분을 담습니다
+//
+//            AlarmService.shared.switchOnOff["total"] = true
+//            AlarmService.shared.switchOnOff["morning"] = true
+//
+//            tableView.reloadData()
+//
+//        }else if cellIdentificationNumber == 2{
+//
+//            AlarmService.shared.mealTime["lunch"] = dateString
+//
+//
+//            AlarmService.shared.mealTimeAMPM["lunch"] = AMPMStr
+//
+//            if AlarmService.shared.mealTimeAMPM["lunch"] == "AM" || AlarmService.shared.mealTimeAMPM["lunch"] == "오전" {
+//                if hourString == "12"{
+//                    AlarmService.shared.mealTimeHour["lunch"] = 0 //오전12시일경우 매칭할 시간이 0시가 됩니다 그래서 0으로 바꿔줍니다
+//                }else{
+//                    AlarmService.shared.mealTimeHour["lunch"] = Int(hourString)
+//                }
+//            }else if AlarmService.shared.mealTimeAMPM["lunch"] == "PM" || AlarmService.shared.mealTimeAMPM["lunch"] == "오후"{
+//                if hourString == "12"{
+//                    AlarmService.shared.mealTimeHour["lunch"] = Int(hourString)
+//                }else{
+//                    AlarmService.shared.mealTimeHour["lunch"] = Int(hourString)! + 12 //오후 1시일 경우 매칭할 숫자가 13시입니다 그래서 12를 더해 담습니다
+//                }
+//
+//            }
+//
+//            AlarmService.shared.mealTimeMinute["lunch"] = Int(minuteStr)
+//
+//            AlarmService.shared.switchOnOff["total"] = true
+//            AlarmService.shared.switchOnOff["lunch"] = true
+//
+//            tableView.reloadData()
+//
+//        }else if cellIdentificationNumber == 3{
+//
+//            AlarmService.shared.mealTime["dinner"] = dateString
+//
+//            AlarmService.shared.mealTimeAMPM["dinner"] = AMPMStr
+//
+//            if AlarmService.shared.mealTimeAMPM["dinner"] == "AM" || AlarmService.shared.mealTimeAMPM["dinner"] == "오전" {
+//                if hourString == "12"{
+//                    AlarmService.shared.mealTimeHour["dinner"] = 0 //오전12시일경우 매칭할 시간이 0시가 됩니다 그래서 0으로 바꿔줍니다
+//                }else{
+//                    AlarmService.shared.mealTimeHour["dinner"] = Int(hourString)
+//                }
+//            }else if AlarmService.shared.mealTimeAMPM["dinner"] == "PM" || AlarmService.shared.mealTimeAMPM["dinner"] == "오후"{
+//                if hourString == "12"{
+//                    AlarmService.shared.mealTimeHour["dinner"] = Int(hourString)
+//                }else{
+//                    AlarmService.shared.mealTimeHour["dinner"] = Int(hourString)! + 12 //오후 1시일 경우 매칭할 숫자가 13시입니다 그래서 12를 더해 담습니다
+//                }
+//
+//            }
+//
+//            AlarmService.shared.mealTimeMinute["dinner"] = Int(minuteStr)
+//
+//            AlarmService.shared.switchOnOff["total"] = true
+//            AlarmService.shared.switchOnOff["dinner"] = true
+//
+//            tableView.reloadData()
+//
+//        }
+//
+//        UserDefaults.standard.setValue(AlarmService.shared.mealTime, forKey: userDefaultsName.mealTime)
+//        UserDefaults.standard.setValue(AlarmService.shared.mealTimeHour, forKey: userDefaultsName.mealTimeHour)
+//        UserDefaults.standard.setValue(AlarmService.shared.mealTimeMinute, forKey: userDefaultsName.mealTimeMinute)
+//        UserDefaults.standard.setValue(AlarmService.shared.mealTimeAMPM, forKey: userDefaultsName.mealTimeAMPM)
+//        UserDefaults.standard.setValue(AlarmService.shared.switchOnOff, forKey: userDefaultsName.alarmOnOff)
+//
+//        print(UserDefaults.standard.dictionary(forKey: userDefaultsName.alarmOnOff) ?? "알람OnOff값이 없음")
+//        print(AlarmService.shared.mealTimeAMPM)
+//
+//        setMealTimeAlarmNotification()
+//
+//        timePickerViewOut.isHidden = true
+//        timePickConfirmBtnOut.isHidden = true
+//
+//
     }
     
 }
